% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-networks.R
\name{xforce_as_networks}
\alias{xforce_as_networks}
\title{Get Networks Assigned to an Autonomous System Number}
\usage{
xforce_as_networks(asn, api_key = Sys.getenv("XFORCE_API_KEY"),
  api_password = Sys.getenv("XFORCE_API_PASSWORD"))
}
\arguments{
\item{asn}{Autonomous System Number}

\item{api_key, api_password}{IBM X-Force API Key & Password. All \code{xforce}
API functions will look for these in the \code{XFORCE_API_KEY} and
\code{XFORCE_API_PASSWORD} environment variables. You can store these
in \code{~/.Renviron} and you can obtain them
\href{https://exchange.xforce.ibmcloud.com/settings/api}{on the IBM X-Force Portal}.}
}
\description{
Get Networks Assigned to an Autonomous System Number
}
\examples{
\\donrun{
xforce_as_networks("3131")
}
}
