% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip-history.R
\name{xforce_ip_history}
\alias{xforce_ip_history}
\title{Get IP History}
\usage{
xforce_ip_history(ip, api_key = Sys.getenv("XFORCE_API_KEY"),
  api_password = Sys.getenv("XFORCE_API_PASSWORD"))
}
\arguments{
\item{ip}{IP address}

\item{api_key, api_password}{IBM X-Force API Key & Password. All \code{xforce}
API functions will look for these in the \code{XFORCE_API_KEY} and
\code{XFORCE_API_PASSWORD} environment variables. You can store these
in \code{~/.Renviron} and you can obtain them
\href{api_key=Sys.getenv("XFORCE_API_KEY")}{on the IBM X-Force Portal}.}
}
\description{
Get IP History
}
\examples{
\\donrun{
xforce_ip_history("174.62.167.97")
}
}
