% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malware-hash.R
\name{xforce_malware_for_hash}
\alias{xforce_malware_for_hash}
\title{Get Malware for File Hash}
\usage{
xforce_malware_for_hash(hash, api_key = Sys.getenv("XFORCE_API_KEY"),
  api_password = Sys.getenv("XFORCE_API_PASSWORD"))
}
\arguments{
\item{hash}{file hash}

\item{api_key, api_password}{IBM X-Force API Key & Password. All \code{xforce}
API functions will look for these in the \code{XFORCE_API_KEY} and
\code{XFORCE_API_PASSWORD} environment variables. You can store these
in \code{~/.Renviron} and you can obtain them
\href{https://exchange.xforce.ibmcloud.com/settings/api}{on the IBM X-Force Portal}.}
}
\description{
Get Malware for File Hash
}
\examples{
\\donrun{
xforce_malware_for_hash("474B9CCF5AB9D72CA8A333889BBB34F0")
}
}
