% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns.R
\name{xforce_resolve}
\alias{xforce_resolve}
\title{Get DNS Records}
\usage{
xforce_resolve(query, api_key = Sys.getenv("XFORCE_API_KEY"),
  api_password = Sys.getenv("XFORCE_API_PASSWORD"))
}
\arguments{
\item{query}{IP address, domain name, or URL}

\item{api_key, api_password}{IBM X-Force API Key & Password. All \code{xforce}
API functions will look for these in the \code{XFORCE_API_KEY} and
\code{XFORCE_API_PASSWORD} environment variables. You can store these
in \code{~/.Renviron} and you can obtain them
\href{https://exchange.xforce.ibmcloud.com/settings/api}{on the IBM X-Force Portal}.}
}
\description{
Get DNS Records
}
\examples{
\\donrun{
xforce_resolve("174.62.167.97")
}
}
