% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url-malware.R
\name{xforce_url_malware}
\alias{xforce_url_malware}
\title{Get Malware for URL}
\usage{
xforce_url_malware(url, api_key = Sys.getenv("XFORCE_API_KEY"),
  api_password = Sys.getenv("XFORCE_API_PASSWORD"))
}
\arguments{
\item{url}{a URL}

\item{api_key, api_password}{IBM X-Force API Key & Password. All \code{xforce}
API functions will look for these in the \code{XFORCE_API_KEY} and
\code{XFORCE_API_PASSWORD} environment variables. You can store these
in \code{~/.Renviron} and you can obtain them
\href{https://exchange.xforce.ibmcloud.com/settings/api}{on the IBM X-Force Portal}.}
}
\description{
Get Malware for URL
}
\examples{
\\donrun{
xforce_url_malware("mediaget.com")
}
}
