% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-profile.R
\name{xforce_user_profile}
\alias{xforce_user_profile}
\title{Get User Profile}
\usage{
xforce_user_profile(api_key = Sys.getenv("XFORCE_API_KEY"),
  api_password = Sys.getenv("XFORCE_API_PASSWORD"))
}
\arguments{
\item{api_key, api_password}{IBM X-Force API Key & Password. All \code{xforce}
API functions will look for these in the \code{XFORCE_API_KEY} and
\code{XFORCE_API_PASSWORD} environment variables. You can store these
in \code{~/.Renviron} and you can obtain them
\href{https://exchange.xforce.ibmcloud.com/settings/api}{on the IBM X-Force Portal}.}
}
\description{
Get User Profile
}
\examples{
\\donrun{
xforce_user_profile()
}
}
