% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whois.R
\name{xforce_whois}
\alias{xforce_whois}
\title{Get WHOIS Information for a Host}
\usage{
xforce_whois(host, api_key = Sys.getenv("XFORCE_API_KEY"),
  api_password = Sys.getenv("XFORCE_API_PASSWORD"))
}
\arguments{
\item{host}{host (domain) name}

\item{api_key, api_password}{IBM X-Force API Key & Password. All \code{xforce}
API functions will look for these in the \code{XFORCE_API_KEY} and
\code{XFORCE_API_PASSWORD} environment variables. You can store these
in \code{~/.Renviron} and you can obtain them
\href{https://exchange.xforce.ibmcloud.com/settings/api}{on the IBM X-Force Portal}.}
}
\description{
Get WHOIS Information for a Host
}
\examples{
\\donrun{
xforce_ip_report("174.62.167.97")
}
}
