% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xforce-package.R
\docType{package}
\name{xforce}
\alias{xforce}
\alias{xforce-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/xforce}
\item BugReports: \url{https://gitlab.com/hrbrmstr/xforce/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
